PpVisual/requirements.txt
absl-py==0.7.1
alabaster==0.7.11
appdirs==1.4.3
argon2-cffi==20.1.0
asn1crypto==0.24.0
astor==0.7.1
astroid==2.0.4
astropy==3.0.4
atomicwrites==1.2.1
attrs==19.3.0
Automat==0.7.0
Babel==2.6.0
backcall==0.1.0
backports.shutil-get-terminal-size==1.0.0
bcrypt==3.2.0
beautifulsoup4==4.6.3
bitarray==0.8.3
bkcharts==0.2
bleach==2.1.4
blis==0.4.1
bokeh==0.13.0
boto==2.49.0
boto3==1.13.12
botocore==1.16.12
Bottleneck==1.2.1
bs4==0.0.1
cachetools==4.1.0
catalogue==1.0.0
certifi==2019.3.9
cffi==1.11.5
chardet==3.0.4
click==6.7
cloudpickle==0.5.5
colorama==0.3.9
comtypes==1.1.7
constantly==15.1.0
contextlib2==0.5.5
crochet==1.12.0
cryptography==2.9.2
cssselect==1.1.0
cycler==0.10.0
cymem==2.0.3
Cython==0.29.14
cytoolz==0.10.1
dask==0.19.1
dataclasses==0.7
decorator==4.3.0
defusedxml==0.5.0
detect==2019.4.13
distributed==1.23.1
docopt==0.6.2
docutils==0.14
entrypoints==0.2.3
et-xmlfile==1.0.1
fake-useragent==0.1.11
fastcache==1.0.2
fastdtw==0.3.2
filelock==3.0.8
fire==0.3.1
Flask==1.0.2
Flask-Cors==3.0.6
future==0.17.1
gast==0.2.2
gensim==3.6.0
gevent==1.3.6
glob2==0.6
google-auth==1.16.0
google-auth-oauthlib==0.4.1
greenlet==0.4.15
grpcio==1.29.0
h5py==2.8.0
HeapDict==1.0.0
html5lib==1.0.1
hyperlink==18.0.0
icc-rt==2019.0
idna==2.7
imageio==2.4.1
imagesize==1.1.0
importlib-metadata==1.7.0
incremental==17.5.0
intel-openmp==2019.0
ipykernel==4.10.0
ipython==6.5.0
ipython-genutils==0.2.0
ipywidgets==7.4.1
isort==4.3.4
itemadapter==0.1.0
itemloaders==1.0.2
itsdangerous==0.24
jange==0.1.6
jdcal==1.4
jedi==0.12.1
jellyfish==0.8.2
jieba==0.42.1
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.13.0
jsonschema==3.2.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
keyring==13.2.1
kiwisolver==1.0.1
langdetect==1.0.7
lazy-object-proxy==1.3.1
llvmlite==0.24.0
locket==0.2.0
lxml==4.2.5
Markdown==3.1
MarkupSafe==1.1.1
matplotlib==2.2.3
mccabe==0.6.1
mistune==0.8.3
mkl==2019.0
mkl-fft==1.0.6
mkl-random==1.0.1
mock==2.0.0
more-itertools==8.4.0
mpmath==1.0.0
msgpack==0.5.6
multipledispatch==0.6.0
murmurhash==1.0.2
mysql-connector==2.2.9
nbconvert==5.6.1
nbformat==4.4.0
networkx==2.4
nltk==3.3
nose==1.3.7
notebook==6.1.1
numba==0.39.0
numexpr==2.6.8
numpy==1.19.0rc1
oauthlib==3.1.0
olefile==0.46
openpyxl==2.5.6
packaging==17.1
pandas==1.0.5
pandocfilters==1.4.2
paramiko==2.7.2
parsel==1.6.0
parso==0.3.1
partd==0.3.8
path.py==11.1.0
pathlib2==2.3.2
patsy==0.5.0
pbr==5.2.0
pep8==1.7.1
pickleshare==0.7.4
Pillow==5.2.0
pipreqs==0.4.10
pkginfo==1.4.2
plac==1.1.3
plotly==4.9.0
pluggy==0.7.1
ply==3.11
preshed==3.0.2
prometheus-client==0.3.1
prompt-toolkit==1.0.15
Protego==0.1.16
protobuf==3.11.3
psutil==5.4.7
psycopg2-binary==2.8.5
py==1.6.0
pyasn1==0.4.4
pyasn1-modules==0.2.2
pybind11==2.2.4
pycodestyle==2.4.0
pycosat==0.6.3
pycparser==2.18
pycurl==7.43.0.2
PyDispatcher==2.0.5
pyflakes==2.0.0
Pygments==2.2.0
PyHamcrest==2.0.2
pylint==2.1.1
PyMySQL==0.9.3
PyNaCl==1.4.0
pyodbc==4.0.24
pyOpenSSL==19.1.0
pyparsing==2.2.0
pyrsistent==0.16.0
PySocks==1.6.8
pytest==3.8.0
pytest-arraydiff==0.2
pytest-astropy==0.4.0
pytest-doctestplus==0.1.3
pytest-openfiles==0.3.0
pytest-remotedata==0.3.0
python-dateutil==2.7.3
pytz==2018.5
PyWavelets==1.0.0
pywin32==223
pywin32-ctypes==0.2.0
pywinpty==0.5.4
PyYAML==3.13
pyzmq==17.1.2
QtAwesome==0.4.4
qtconsole==4.4.1
QtPy==1.5.0
queuelib==1.5.0
regex==2020.5.14
requests==2.23.0
requests-oauthlib==1.3.0
retrying==1.3.3
rope==0.11.0
rsa==4.0
ruamel.yaml==0.15.46
s3transfer==0.3.3
sacremoses==0.0.43
scikit-image==0.14.0
scikit-learn==0.23.2
scipy==1.1.0
Scrapy==2.1.0
seaborn==0.9.0
segtok==1.5.0
selenium==3.141.0
Send2Trash==1.5.0
sentencepiece==0.1.91
service-identity==17.0.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.11.0
smart-open==2.0.0
snowballstemmer==1.2.1
snownlp==0.12.3
sortedcollections==1.0.1
sortedcontainers==2.0.5
soupsieve==2.0.1
spacy==2.3.2
sphinxcontrib-websupport==1.1.0
spyder-kernels==0.2.6
SQLAlchemy==1.3.1
srsly==1.0.2
statsmodels==0.9.0
sympy==1.1.1
tables==3.4.4
tabulate==0.8.7
tb-nightly==2.3.0a20200604
tbb==2019.0
tbb4py==2019.0
tblib==1.3.2
tensorboard==1.14.0
tensorboard-plugin-wit==1.6.0.post3
tensorboardX==2.0
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.3.1
thinc==7.4.1
threadpoolctl==2.1.0
tokenizers==0.7.0
toolz==0.9.0
torch==1.5.0+cpu
torchvision==0.6.0+cpu
tornado==5.1
tqdm==4.46.1
traitlets==4.3.2
transformers==2.11.0
Twisted==18.7.0
typed-ast==1.4.1
unicodecsv==0.14.1
urllib3==1.23
urlopen==1.0.0
w3lib==1.22.0
wasabi==0.7.1
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.4.2
win-inet-pton==1.0.1
win-unicode-console==0.5
wincertstore==0.2
wrapt==1.10.11
xlrd==1.1.0
XlsxWriter==1.1.0
xlwings==0.11.8
xlwt==1.3.0
yake @ git+https://github.com/LIAAD/yake@2c9fedbe0e3195a273a3c3daef3d41d4d7c5b3b3
yarg==0.1.9
zict==0.1.3
zipp==3.1.0
zope.interface==4.5.0
